requirements.txt
absl-py==0.9.0
astor==0.8.1
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
importlib-metadata==1.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
llvmlite==0.31.0
Markdown==3.2.2
numba==0.47.0
numpy==1.18.5
protobuf==3.12.2
python-speech-features==0.6
resampy==0.2.2
scipy==1.4.1
six==1.15.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
